refactor: 增强安全防护并清理冗余代码
- security(fs): 添加二进制文件检测和工作区边界检查函数
  - 新增 is_binary_file() 函数，通过扩展名和内容采样检测二进制文件
  - 新增 check_workspace_boundary() 函数，防止路径遍历攻击
- security(http): 添加 HTTP 头注入防护
  - 验证 header 名称和值是否符合 RFC 7230 规范
  - 拒绝包含 CRLF、NUL、DEL 等非法控制字符的请求头
- refactor(storage): 移除冗余的迁移记录函数
  - 移除 record_migration() 和 remove_migration()，逻辑已内联到 Database::migrate()
- refactor(logging): 统一使用 spdlog 接口
  - migration 模块从 TURBOT_LOG_* 宏改为直接使用 spdlog
- refactor(tools): 优化文件工具的错误处理和边界检查
  - read_file_tool 使用共享的 is_binary_file() 检测
  - bash_tool/glob_tool/list_tool 应用工作区边界验证
- refactor(utils): 优化字符串工具和环境工具实现
  - string_utils: 改进字符串分割和修剪逻辑
  - env_utils: 增强环境变量处理
- build: 更新 CMakeLists.txt 依赖配置

+++ libs/utils/src/crypto_utils.cpp
@@ -358,7 +358,8 @@ std::string aes_256_gcm_decrypt(const AesGcmResult& encrypted, const std::string
 
     // 设置tag - OpenSSL API requires non-const pointer
     std::vector<uint8_t> tag_copy(encrypted.tag.begin(), encrypted.tag.end());
-    if (EVP_CIPHER_CTX_ctrl(ctx.get(), EVP_CTRL_GCM_SET_TAG, encrypted.tag.size(),
+    if (EVP_CIPHER_CTX_ctrl(ctx.get(), EVP_CTRL_GCM_SET_TAG,
+                            static_cast<int>(encrypted.tag.size()),
                             tag_copy.data()) != 1) {
         throw std::runtime_error("Failed to set tag");
     }
